Move levenshtein distance to StringUtil
incorrect classification under CollectionUtil
