+++ src/main/java/seedu/address/model/person/Person.java
@@ -1,8 +1,8 @@
 package seedu.address.model.person;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.util.CollectionUtil.levenshteinDistance;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
+import static seedu.address.commons.util.StringUtil.levenshteinDistance;
 
 import java.util.Collections;
 import java.util.List;
